pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pysigrok_vcd"
authors = [{name = "Scott Shawcroft", email = "scott@tannewt.org"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "pysigrok >= 0.4.0",
    "pyvcd < 0.4.0"
]
requires-python = ">=3.7"

[project.urls]
Home = "https://github.com/pysigrok/format-vcd"

[project.entry-points."pysigrok.output_format"]
vcd = "pysigrok_vcd:VCDOutput"
